requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
boto3==1.43.6
botocore==1.43.6
brotli==1.2.0
cbor2==6.0.1
celery==5.6.3
certifi==2026.4.22
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==48.0.0
cssselect2==0.9.0
daphne==4.2.1
Django==4.2.30
django-allauth==65.16.1
django-ckeditor==6.7.3
django-debug-toolbar==6.3.0
django-js-asset==3.1.2
django-storages==1.14.6
fonttools==4.62.1
gunicorn==26.0.0
hyperlink==21.0.0
idna==3.14
Incremental==24.11.0
jmespath==1.1.0
kombu==5.6.2
msgpack==1.1.2
packaging==26.2
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydyf==0.12.1
PyJWT==2.12.1
pyOpenSSL==26.2.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
razorpay==2.0.1
redis==7.4.0
requests==2.33.1
s3transfer==0.17.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.1.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
ujson==5.12.1
urllib3==2.7.0
vine==5.1.0
wcwidth==0.7.0
weasyprint==68.1
webencodings==0.5.1
whitenoise==6.12.0
zope.interface==8.4
zopfli==0.4.1
